requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
twilio==8.11.0
gunicorn==21.2.0
flask-talisman==1.1.0
stripe==7.8.0
supabase==2.3.0
